Alteração de compatibidade
Inclusão da notação "@Id" para identificar a chave primária e alteração do tipo de dado para Long no atributo chave.

+++ src/br/ufac/academico/entity/Atendente.java
@@ -5,7 +5,8 @@ import javax.persistence.Entity;
 @Entity
 public class Atendente {
 	
-	private int codigo;
+	
+	private long codigo;
 	private String nome;
 	private String ultimoAcesso;
 	private String ramal;
